Rename `userEvent` to `fireUserEvent` to avoid confusing the `testing-library` eslint plugin
